# api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "用户管理API",
    "version": "1.0.0",
    "description": "用户管理系统API接口文档"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "开发环境"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "用户登录",
        "tags": ["认证"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "用户名",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "description": "密码",
                    "example": "admin123"
                  },
                  "remember_me": {
                    "type": "boolean",
                    "description": "记住我",
                    "example": false
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "description": "访问令牌"
                    },
                    "token_type": {
                      "type": "string",
                      "description": "令牌类型",
                      "example": "bearer"
                    },
                    "expires_in": {
                      "type": "integer",
                      "description": "过期时间（秒）",
                      "example": 1800
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "description": "用户ID"
                        },
                        "username": {
                          "type": "string",
                          "description": "用户名"
                        },
                        "full_name": {
                          "type": "string",
                          "description": "全名"
                        },
                        "role": {
                          "type": "string",
                          "description": "角色"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误"
          },
          "423": {
            "description": "账户被锁定"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "获取用户列表",
        "tags": ["用户管理"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "username": {
                            "type": "string"
                          },
                          "full_name": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "size": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建用户",
        "tags": ["用户管理"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "用户名"
                  },
                  "password": {
                    "type": "string",
                    "description": "密码"
                  },
                  "full_name": {
                    "type": "string",
                    "description": "全名"
                  },
                  "role": {
                    "type": "string",
                    "description": "角色",
                    "enum": ["admin", "project_manager", "tester", "developer"]
                  }
                },
                "required": ["username", "password", "full_name", "role"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          },
          "400": {
            "description": "参数错误"
          },
          "409": {
            "description": "用户名已存在"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}